modification view

+++ view/Repliques/detail.php
@@ -1,13 +1,16 @@
 <?php
-echo htmlspecialchars($r->get('idReplique'));
-echo htmlspecialchars(' ' . $r->get('nomReplique'));
-echo htmlspecialchars(' ' . $r->get('nomCategorie'));
-echo htmlspecialchars(' ' . $r->get('stockRepliques') . ' ');
+$idReplique = htmlspecialchars($r->get('idReplique'));
+$nomReplique = htmlspecialchars($r->get('nomReplique'));
+$nomCategorie = htmlspecialchars($r->get('nomCategorie'));
+$stockReplique = htmlspecialchars($r->get('stockRepliques'));
 $rawUrlReplique = rawurlencode($r->get('idReplique'));
-var_dump($_SESSION);
-    echo('<a href="?controller=repliques&action=delete&idReplique=' . $rawUrlReplique . '"     >Supprimer</a>');
-    echo '<br>';
-    echo('<a href="?controller=repliques&action=update&idReplique=' . $rawUrlReplique . '"     >Modifier</a>');
-    echo '<br>';
 
-echo('<a href="?controller=panier&action=set&idReplique=' . $rawUrlReplique . '&qte=1"     >Ajouter au panier</a>');
+echo "<p>$nomReplique de catégorie $nomCategorie ($stockReplique en stock) ";
+echo('<a href="?controller=panier&action=set&idReplique=' . $rawUrlReplique . '&qte=1">Ajouter au panier </a>');
+
+if (isset($_SESSION['codeClient']) && $_SESSION['admin']) {
+    echo('<a href="?controller=repliques&action=delete&idReplique=' . $rawUrlReplique . '">Supprimer </a>');
+    echo('<a href="?controller=repliques&action=update&idReplique=' . $rawUrlReplique . '">Modifier </a>');
+}
+
+echo "</p>";
